requirements.txt
fastapi>=0.100.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
uvicorn>=0.23.0
python-dotenv>=1.0.0
google-generativeai>=0.4.0
requests>=2.0.0
httpx>=0.24.1
motor>=3.3.0
pymongo>=4.5.0
web3>=6.10.0
eth-account>=0.10.0
pyjwt>=2.8.0
solana>=0.30.2
anchorpy>=0.17.1
base58>=2.1.1

# Testing dependencies
pytest>=7.3.1
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-xdist>=3.3.1

# Development tools
black
flake8
isort
ccxt>=4.0.0
firecrawl-py>=0.0.7 